+++ ESFramework/ESMacros.h
@@ -39,13 +39,6 @@
 #define ESStrongSelf    ESStrong(self); if (!_self) return;
 
 /**
- * Bits-mask helper.
- */
-#define ESMaskIsSet(value, flag)    (((value) & (flag)) == (flag))
-#define ESMaskSet(value, flag)      ((value) |= (flag));
-#define ESMaskUnset(value, flag)    ((value) &= ~(flag));
-
-/**
  * Defines a key for the associcated object.
  */
 #define ESDefineAssociatedObjectKey(name) static const void * name##Key = &name##Key;